requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
email-validator
requests
Pillow
prometheus-flask-exporter




